requirements.txt
numpy==2.1.3
winrt-Windows.UI==3.0.0
winrt-Windows.UI.ViewManagement==3.0.0
faster-whisper==1.1.1
PyQt6-Fluent-Widgets[full]==1.7.6 
PyQt6==6.8.1
psutil==7.0.0
nvidia-cuda-runtime-cu12==12.6.77,
nvidia-cublas-cu12==12.6.4.1,
nvidia-cudnn-cu12==9.5.1.17